Demo/shopTHAT/backend/parser/pyproject.toml
[tool.poetry]
name        = "pyparser-toml"
version     = "0.1.0"
description = ""
authors     = ["Lokesh Katuri"]
packages    = [{ include = "pyparser" }]

[tool.poetry.dependencies]
python                  = ">=3.10,<4.0"
beautifulsoup4          = "4.13.4"
bs4                     = "0.0.2"
faiss-cpu               = "1.11.0"
sentence-transformers   = "3.4.1"
yake                    = "0.4.8"
numpy                   = "2.2.0"
requests                = "2.32.3"
groq                    = "0.23.1"
tenacity                = "9.1.2"
pandas                  = "2.2.3"
openpyxl                = "3.1.5"
selenium                = "4.31.0"
webdriver-manager       = "4.0.2"
langchain               = "0.3.24"
langchain-community     = "0.3.23"
unstructured            = "0.17.2"
scrapy                  = "2.12.0"
langchain-huggingface   = "0.1.2"
mlflow                  = "2.22.0"
pydantic                = "2.11.4"
pydantic-settings       = "2.9.1"
fastapi                 = "*"   # or pin to e.g. "0.100.0" if you need a specific version
torchvision             = "*"
pillow-avif-plugin      = "1.5.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]


[build-system]
requires    = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"